requirements.txt
requests
bs4
tqdm
pandas
lyricsgenius
openai
seaborn
plotly
matplotlib
tueplots
langdetect
scikit-learn
tenacity
jaro-winkler
jupyter
python-dotenv
httpx
openpyxl
lxml
orjson
unidecode
torch
transformers
scikit-learn
wandb
accelerate
datasets
evaluate
numpy
